Removing option to generate extended GO Annotation tables
It was buggy. At least using the default EBI Gene-Ontology MySQL instance.
The problem was that queries did not accept all Terms submitted (if too many)
and then did not return parent terms for all original queries. Hence we deprecate
this function. Possibly works with local installations. But better would be
to use the OBO OWL files with adequate Java libraries to parse these.