Improve restrictions on eval-output of max attainable GO annots scores
Bring the conditional structure of the GO annotation score related
output columns in the EvaluatorOutputWriter in line with the structure
controlling the column heads.
This avoids situations where the column heads are miss matched with the
data columns and makes sure the data to be printed has actually been
calculated.