Default to filtering the ground truth HRD with the token black list
If the evaluateOnlyValidTokens parameter is set to true (its new default
since the last commit) filtering is to be applied to the ground truth
description too (except if the groundTruthTokenBlackList is set).
Empty ground truth descriptions can emerge when all tokens are found to
have a match in the token black list (eg "Protein ORF3").
In such cases AHRD now gets a description score of 1 if no annotation
was made. For fairness reasons the competitors have to be get the same
score in such cases. Thus an empty best blast result is given to each
blast database in order to have something to compare if there aren't any
blast results for the particular blast database. In the vast majority of
cases will either a real blast result for the database be found which
will then replace the empty blast result or the ground truth description
won't be empty resulting in the appropriate score of 0 for the empty
blast result.
Analogously change AHRD's behavior in regards to competitor annotations
(descriptions as well as GO terms).


Refactor Settings.isInTrainingMode() to Settings.isInEvaluationMode()
because this is its de facto purpose. The former name was misleading.

Setting the writeBestBlastHitsToOutput parameter outside of the
evaluation mode makes no sense but was previously possible and led to
exceptions. It is thus now restricted to evaluation mode.