Set Recall/Precision of empty GroundTruth/Prediction to NaN instead of 1
- Set the Fscore to 0 only if both precision and recall are 0 (No edge
case, just bad prediction). Arithmetically this case should result in
NaN but it should be distinguishable from the edge cases. If only one of
precision or recall is 0 the calculation of the Fscore can proceed and
will correctly result in 0. If either one in NaN the calculation will
result in NaN as well.
- Modify all Fscore test methods to expect NaN values in the relevant
edge cases.
- Exclude NaN values from the calculation of average scores in the
training